# docs/api_swagger.yaml
basePath: /api
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  schema.Account:
    properties:
      application:
        $ref: '#/definitions/schema.Application'
      application_id:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      roleApplications:
        items:
          $ref: '#/definitions/schema.RoleApplication'
        type: array
      status:
        type: boolean
      updatedAt:
        type: string
      user_code:
        type: string
    type: object
  schema.ActivityLog:
    properties:
      application:
        type: string
      body:
        type: string
      client:
        type: string
      createdAt:
        type: string
      duration:
        type: string
      id:
        type: integer
      method:
        type: string
      path:
        type: string
      path_op:
        type: string
      referrer:
        type: string
      request_id:
        type: string
      source:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/schema.User'
      user_id:
        type: string
    type: object
  schema.Application:
    properties:
      accounts:
        items:
          $ref: '#/definitions/schema.Account'
        type: array
      app_package:
        type: string
      app_package_class:
        type: string
      asset_apk:
        type: string
      asset_icon:
        type: string
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updated_note:
        type: string
      updatedAt:
        type: string
      version:
        type: string
    type: object
  schema.ApplicationResponse:
    properties:
      app_package:
        type: string
      app_package_class:
        type: string
      asset_apk:
        type: string
      asset_icon:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_note:
        type: string
      version:
        type: string
    type: object
  schema.Article:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      status:
        type: boolean
      title:
        type: string
      updatedAt:
        type: string
      user_id:
        type: string
    type: object
  schema.Company:
    properties:
      address:
        type: string
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      domain:
        type: string
      fax:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sector:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
  schema.CompanyResponse:
    properties:
      address:
        type: string
      code:
        type: string
      description:
        type: string
      domain:
        type: string
      employees:
        items:
          $ref: '#/definitions/schema.Employee'
        type: array
      estates:
        items:
          $ref: '#/definitions/schema.Estate'
        type: array
      fax:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sector:
        type: string
    required:
    - address
    - code
    - description
    - domain
    - fax
    - name
    - phone_number
    - sector
    type: object
  schema.Division:
    properties:
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      estate:
        $ref: '#/definitions/schema.Estate'
      estate_code:
        type: string
      gangs:
        items:
          $ref: '#/definitions/schema.Gang'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
  schema.DivisionResponse:
    properties:
      code:
        type: string
      delete:
        type: boolean
      description:
        type: string
      estate_id:
        type: string
      gang:
        items:
          $ref: '#/definitions/schema.Gang'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
    type: object
  schema.Employee:
    properties:
      address:
        type: string
      code:
        type: string
      company:
        $ref: '#/definitions/schema.Company'
      company_code:
        type: string
      cost_center:
        type: string
      created_by:
        type: string
      createdAt:
        type: string
      deleted:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      department:
        type: string
      email:
        type: string
      expired_time:
        type: integer
      expiredAt:
        type: string
      fiscal_period:
        type: string
      fiscal_year:
        type: string
      fullname:
        type: string
      grade_rate:
        type: string
      harvester_code:
        type: string
      id:
        type: string
      job_pos:
        type: string
      mandt:
        type: string
      nickname:
        type: string
      nik:
        type: string
      office_number:
        type: string
      phone_number:
        type: string
      picture:
        type: string
      status:
        type: boolean
      type:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/schema.User'
      user_code:
        type: string
      username:
        type: string
    type: object
  schema.Estate:
    properties:
      code:
        type: string
      company:
        $ref: '#/definitions/schema.Company'
      company_code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      divisions:
        items:
          $ref: '#/definitions/schema.Division'
        type: array
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
  schema.EstateResponse:
    properties:
      Division:
        items:
          $ref: '#/definitions/schema.Division'
        type: array
      code:
        type: string
      company:
        $ref: '#/definitions/schema.Company'
      delete:
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
    type: object
  schema.Gang:
    properties:
      code:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      division:
        $ref: '#/definitions/schema.Division'
      division_code:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
    type: object
...
  schema.MsgResponse:
    properties:
      msg:
        type: string
    type: object
  schema.RequestApplication:
    properties:
      app_package:
        type: string
      app_package_class:
        type: string
      asset_apk:
        type: string
      asset_icon:
        type: string
      description:
        type: string
      name:
        type: string
      updated_note:
        type: string
      version:
        type: string
    required:
    - app_package
    - app_package_class
    - asset_apk
    - asset_icon
    - description
    - name
    - updated_note
    - version
    type: object
  schema.RequestApplicationUpdate:
    properties:
      app_package:
        type: string
      app_package_class:
        type: string
      asset_apk:
        type: string
      asset_icon:
        type: string
      delete:
        type: boolean
      description:
        type: string
      name:
        type: string
      status:
        type: boolean
      updated_note:
        type: string
      version:
        type: string
    required:
    - app_package
    - app_package_class
    - asset_apk
    - asset_icon
    - delete
    - description
    - name
    - status
    - updated_note
    - version
    type: object
  schema.RequestCompany:
    properties:
      address:
        type: string
      code:
        type: string
      description:
        type: string
      domain:
        type: string
      fax:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sector:
        type: string
    required:
    - address
    - code
    - description
    - domain
    - fax
    - name
    - phone_number
    - sector
    type: object
  schema.RequestCompanyUpdate:
    properties:
      address:
        type: string
      code:
        type: string
      delete:
        type: boolean
      description:
        type: string
      domain:
        type: string
      fax:
        type: string
      name:
        type: string
      phone_number:
        type: string
      sector:
        type: string
      status:
        type: boolean
    required:
    - address
    - code
    - delete
    - description
    - domain
    - fax
    - name
    - phone_number
    - sector
    - status
    type: object
  schema.RequestDivision:
    properties:
      code:
        type: string
      description:
        type: string
      estate_code:
        type: string
      name:
        type: string
    required:
    - code
    - description
    - name
    type: object
  schema.RequestDivisionUpdate:
    properties:
      code:
        type: string
      delete:
        type: boolean
      description:
        type: string
      estate_code:
        type: string
      name:
        type: string
      status:
        type: boolean
    required:
    - code
    - delete
    - description
    - name
    - status
    type: object
  schema.RequestEstate:
    properties:
      code:
        type: string
      company_code:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - code
    - company_code
    - description
    - name
    type: object
...
  schema.Role:
    properties:
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
  schema.RoleApplication:
    properties:
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  schema.User:
    properties:
      accounts:
        items:
          $ref: '#/definitions/schema.Account'
        type: array
      activityLogs:
        items:
          $ref: '#/definitions/schema.ActivityLog'
        type: array
      articles:
        items:
          $ref: '#/definitions/schema.Article'
        type: array
      code:
        type: string
      createdAt:
        type: string
      delete:
        type: boolean
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      divisions:
        items:
          $ref: '#/definitions/schema.Division'
        type: array
      email:
        type: string
      employees:
        items:
          $ref: '#/definitions/schema.Employee'
        type: array
      id:
        type: string
      roles:
        items:
          $ref: '#/definitions/schema.Role'
        type: array
      status:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Fajar soegi
    url: http://www.swagger.io/support
  description: This is User Service Test.
  termsOfService: http://swagger.io/terms/
  title: Swagger User Service
  version: "1.0"
paths:
  /application:
    get:
      consumes:
      - application/json
      description: Get Application All
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.ApplicationResponse'
            type: array
      summary: Get All Application
      tags:
      - (E) Application
    patch:
      consumes:
      - application/json
      description: Update Application
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Application Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestApplicationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Update Application
      tags:
      - (E) Application
    post:
      consumes:
      - application/json
      description: Create Application
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Estate Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestApplication'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Create Application
      tags:
      - (E) Application
  /application/{id}:
    get:
      consumes:
      - application/json
      description: Get Application ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Application ID OR Application Code
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.ApplicationResponse'
      summary: Get Application By ID
      tags:
      - (E) Application
  /company:
    get:
      consumes:
      - application/json
      description: Get Company All
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.CompanyResponse'
            type: array
      summary: Get All Company
      tags:
      - (A) Company
    post:
      consumes:
      - application/json
      description: Create Company
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Estate Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Create Company
      tags:
      - (A) Company
  /company/{id}:
    get:
      consumes:
      - application/json
      description: Get Company ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Company ID OR Company Code
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.CompanyResponse'
      summary: Get Company By ID
      tags:
      - (A) Company
  /division:
    get:
      consumes:
      - application/json
      description: Get Division All
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.DivisionResponse'
            type: array
      summary: Get All Division
      tags:
      - (C) Division
    patch:
      consumes:
      - application/json
      description: Update Division
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Company Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestDivisionUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Update Division
      tags:
      - (C) Division
    post:
      consumes:
      - application/json
      description: Create Division
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Estate Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestDivision'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Create Division
      tags:
      - (C) Division
  /division/{id}:
    get:
      consumes:
      - application/json
      description: Get Division ID
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Division ID OR Division Code
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.DivisionResponse'
      summary: Get Division By ID
      tags:
      - (C) Division
...
  /estate:
    get:
      consumes:
      - application/json
      description: Get Estate
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schema.EstateResponse'
            type: array
      summary: Get Estate
      tags:
      - (B) Estate
    patch:
      consumes:
      - application/json
      description: Update Company
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Company Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestCompanyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Update Company
      tags:
      - (A) Company
    post:
      consumes:
      - application/json
      description: Create Estate
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 'Estate Schema '
        in: body
        name: Estate
        required: true
        schema:
          $ref: '#/definitions/schema.RequestEstate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.MsgResponse'
      summary: Create Estate
      tags:
      - (B) Estate
...
swagger: "2.0"